feat(payment-service): microservicio de pagos
- Se integro Stripe como pasarela de pagos
- Se desarrollo la logica para las transacciones y pagos
- Se agregaron cliente feign en la libreria commons
- Se agrego cliente feign de notificaciones
- Se modifico clientes feign de pacientes y security
- Se agregaron perfiles de entorno para microservicio de pagos

+++ healthyme-notification/src/main/java/studio/devbyjose/healthyme_notification/event/CitaEvent.java
@@ -15,7 +15,7 @@ import java.time.LocalTime;
 public class CitaEvent {
     private Integer idCita;
     private TipoEventoCita tipoEvento;
-    private Integer idPaciente;
+    private Long idPaciente;
     private String emailPaciente;
     private Integer idMedico;
     private String especialidad;
